Change stat values from double to int

+++ TurnBasedRPG/Enemies/Monster.cs
@@ -3,10 +3,10 @@
     internal class Monster : IMonster
     {
         public Monster(
-            double health,
-            double physicDefense,
-            double magicDefense,
-            double strength,
+            int health,
+            int physicDefense,
+            int magicDefense,
+            int strength,
             IEnumerable<Skill> skills,
             int experienceToDrop,
             EnemyTypes type)
@@ -20,10 +20,10 @@
             Type = type;
         }
 
-        public double Health { get; set; }
-        public double Armor { get; set; }
-        public double MagicDefense { get; set; }
-        public double Strength { get; set; }
+        public int Health { get; set; }
+        public int Armor { get; set; }
+        public int MagicDefense { get; set; }
+        public int Strength { get; set; }
 
         public IEnumerable<Skill> Skills { get; }
